Maintain SPI RX order while byte is being received
Before, it was possible for the core to change the SPI configuration in
the middle of communication, even though the docs say not to do this.
This violates output determinism in the case of the SPI-in clock domain,
so we patch the peripheral to prevent changing the RX order config while
in the middle of receiving a byte. This doesn't fix the
CDC/metastability problems with changing config, but it does fix the
security issue.