Fix USB reset sync to satisfy output determinism
The original USB synchronizer design violates output determinism
(although likely not in a way that could leak secrets in any real world
application). Note that this change is reflected separately in Kronos
(which has its own top level), but we copied it here for benchmarking
purposes.

We chose this new design because it is considered good practice (see
"Asynchronous & Synchronous Reset Design Techniques - Part Deux" by
Cummings et al.), and used in the newer OT reset manager. We also `&`
the output of the synchronizer with the reset signal to ensure we can
get reset everything in one step from Yosys.